add Clock
added new Clock object.

NOTE: requires C++20.

Unlike the Basic Clock, which always uses the internal system clock to determine the time (and also cannot take into account daylight saving times, for example)...

Clock can take into account user specified time zones (as well as automatically correcting for daylight savings).
The "Time" structure used in this Clock has much more information than the one in BasicClock, which just has hour, minute and second.
Clock's Time includes: the full time including milliseconds as well the full date, including day of the week. They are all stored as unsigned ints.
For fields that are not directly equal values:
month: 1 (January) - 12 (December),
day: 1 - 31 (day of the month),
weekday: 0* (Sunday) - 7* (Sunday),
hour: 0 - 23.
Weekday can be either 0 - 6 or 1 - 7, with 0 or 7 being Sunday and 1 - 6 always being being Monday to Saturday.
Whether Sunday is 0 or 7 depends on whether or not ISO Encoding is set (ISO encoding uses 7 for Sunday otherwise it will be 0).

The time zone is set by specifying an official time zone code ("TZ Identifier").
If the specified TZ Identifier is valid and successful, it will use it for the time. If it is not valid or successful, it will use the time zone of the device. And if the time zone of the device fails, it will resort to using the internal system clock (as BasicClock does).
You can get the currently active time zone "TZ Identifier" to check that the one you specified was successful. Note that it may be different from the exact one you specify as there are many aliases. For example, if you specify "Japan", it will likely set the TZ Identifier to "Asia/Tokyo". This means that it was successful but comparing the strings will be unequal.
If the time zones were all successful and the internal system clock is used, the active TZ Identifier is simply "" (an empty string).